+++ jdaviz/components/plugin_plot.vue
@@ -1,7 +1,9 @@
 <template>
-  <div class="plugin-plot-component">
-    <v-row style="margin: 0px 0px -8px 0px !important" justify="end">
-      <div style="line-height: 64px; width=32px" class="only-show-in-tray">
+  <div class="plugin-plot-component" style="margin-bottom: 40px">
+    <v-row dense style="margin: 0px 0px -8px 0px !important">
+      <jupyter-widget class='plugin-nested-toolbar' :widget="toolbar"></jupyter-widget>
+      <v-spacer></v-spacer>
+      <div style="line-height: 40px; width=32px" class="only-show-in-tray">
         <j-plugin-popout :popout_button="popout_button"></j-plugin-popout>
       </div>
     </v-row>
@@ -9,13 +11,11 @@
     <v-row style="margin: -16px 0px 8px 0px !important">
       <jupyter-widget :widget="figure" style="height: 100%; width: 100%" />
     </v-row>
-
   </div>
 </template>
 
 <script>
 </script>
-
 
 <style scoped>
   .only-show-in-tray {
@@ -33,5 +33,4 @@
     width: 100%;
     height: 480px;
   }
-
 </style>