+++ CHANGES.rst
@@ -4,6 +4,9 @@
 New Features
 ------------
 
+- Plots in plugins now include basic zoom/pan tools for Plot Options,
+  Imviz Line Profiles, and Imviz's aperture photometry. [#2498]
+
 Cubeviz
 ^^^^^^^